+++ src/components/ProductList/index.tsx
@@ -1,0 +1,3 @@
+import ProductList from './ProductList';
+
+export default ProductList;
