+++ src/recoil/cartAtoms.ts
@@ -1,6 +1,6 @@
 import { atom } from 'recoil';
 import { Cart } from '../types/types';
-import { recoilPersist } from 'recoil-persist'
+import { recoilPersist } from 'recoil-persist';
 
 const { persistAtom } = recoilPersist({
   key: 'recoil-persist',
